Improve comments. Move LoadPlaybookFolder to playbook module.
