+++ playbook/playbook.go
@@ -2,6 +2,7 @@ package playbook
 
 import (
 	"errors"
+	"fmt"
 	"gopkg.in/yaml.v2"
 	"io/ioutil"
 	"os"
@@ -100,4 +101,30 @@ func ParsePlaybook(playbook []byte) (Playbook, error) {
 
 func ReadPlaybookFromDisk(fd string) ([]byte, error) {
 	return ioutil.ReadFile(fd)
+}
+func LoadPlaybookFolder(dir string) ([]Playbook, error) {
+	var AllPlaybooks []Playbook
+	paths, err := filepath.Glob(dir + "/*")
+	if err != nil {
+		return AllPlaybooks, err
+	}
+	for _, path := range paths {
+		playbookBytes, err := ReadPlaybookFromDisk(path)
+		if err != nil {
+			fmt.Printf("Warning: Failed to read %s\n", path)
+			continue
+		}
+		parsed, err := ParsePlaybook(playbookBytes)
+		if err != nil {
+			fmt.Printf("Warning: Failed to parse %s\n", path)
+			continue
+		}
+		err = parsed.Validate()
+		if err != nil {
+			fmt.Printf("Warning: Playbook %s invalid: %s\n", path, err)
+			continue
+		}
+		AllPlaybooks = append(AllPlaybooks, parsed)
+	}
+	return AllPlaybooks, nil
 }
