+++ playbook/playbook_test.go
@@ -160,6 +160,15 @@ func TestValidatePlaybookFailures(t *testing.T) {
 	}
 }
 
+func TestLoadPlaybookFolder(t *testing.T) {
+	pbs, err := LoadPlaybookFolder("playbooks/")
+	if err != nil {
+		t.Errorf("LoadPlaybookFolder failed to load playbooks: %s\n", err)
+	}
+	if len(pbs) == 0 {
+		t.Error("LoadPlaybookFolder failed to load mock playbook")
+	}
+}
 func TestTaskManifestsPresentPasses(t *testing.T) {
 	testcases := []struct {
 		scenario string
